Overview

Feynman.Feynman

Feynman.ISP

Feynman.balancingIdeal

Feynman.computeBaikovMatrix

Feynman.computeIBP

Feynman.eliminateVariables

Feynman.feynmanDenominators

Feynman.labelGraph

Feynman.makePoly

Feynman.printIBP

Feynman.printLabeledGraph

Feynman.propagators

Feynman.removeElimVars

Feynman.removeParameter

Feynman.removeVariable

Feynman.removeVariableLocal

Feynman.substituteGraph

Feynman.Feynman — Module

Feynman is a package for computing integration-by-part identities (IBPs) of a Feynman Integral associated to Feynman graph using module intesecation method. 
This package also provides an interface of Oscar to use the packages NeatIBP developed using Singular and GPI-Space.

Feynman.ISP — Method

ISP(G::labeledgraph)

USAGE : ISP(G);

ASSUME : G is a labeled graph.

RETURN : idal containing the irreducible scalar products(ISPs), that is, those scalar product which are not linearly dependent on the propagators.

#Examples

julia> G=simple_graph([1,2,3,4,5,6],[(1,2),(3,6),(4,5),(1,6),(2,3),(5,6),(3,4),1,2,5,4]);

julia> G=labelGraph(G,0);

julia> Gelim=eliminateVariables(G);

julia> ISP(Gelim)
Ideal generated by
  p[3]*q[1]
  p[1]*q[2]

Feynman.balancingIdeal — Method

balancingIdeal(G::labeledgraph)

USAGE : balancingIdeal(G);

ASSUME : G is a labeled graph.

RETURN : Ideal of balancing condition of the graph. i.e Ideal generated by the relation of the momentums which are obtained by applying momentum conservation law to external mementa, and at each vertex; This is an ideal of the ring G.over.

#Examples

julia> G=simple_graph([1,2,3,4],[(1,3),(1,2),(2,4),(3,4),1,2,3,4]);
julia> G=labelGraph(G,0);
julia> balancingIdeal(G)

Ideal generated by
p[1] + p[2] + p[3] + p[4]
p[1] + q[1] + q[2]
p[2] - q[2] + q[3]
p[3] - q[1] + q[4]
p[4] - q[3] - q[4]

Feynman.computeBaikovMatrix — Method

computeBaikovMatrix(G::simple_graphgraph)

USAGE : computeBaikovMatrix(G);

ASSUME : G is a graph, or G is a labled graph where redundant variables have been eliminated by the procedure eliminateVariables, and deleted from the ring by the procedure removeElimVars.

RETURN : a labeled graph G, where the computed Baikov matrix and the polynomial ring where baikovmatrix is defined are stored in G.baikovmatrix and G.baikovover respectively.

#Examples

julia> G=simple_graph([1,2,3,4,5,6,7],[(6,1),(6,4),(1,2),(3,7),(4,3),(2,7),(5,6),(7,5),1,2,3,4,5]);

julia> G=Feynman.labelGraph(G,0);

julia> G=Feynman.eliminateVariables(G);

julia> G=Feynman.removeElimVars(G);
QQMPolyRingElem[p[1], p[2], p[3], p[4], p[5], q[1], q[2]]
julia> G=Feynman.computeBaikovMatrix(G);
labels used for Gram matrix of external loop momenta:
["p[1]*p[2] => 1//2*t[1]"]
["p[1]*p[3] => 1//2*t[2]"]
["p[2]*p[3] => 1//2*t[3]"]
["p[1]*p[4] => 1//2*t[4]"]
["p[2]*p[4] => 1//2*t[5]"]
["p[3]*p[4] => -1//2*t[1] - 1//2*t[2] - 1//2*t[3] - 1//2*t[4] - 1//2*t[5]"]
Assignment of Baikov variables (Z_i) are:
["z[1] => p[3]*q[1]"]
["z[2] => p[4]*q[1]"]
["z[3] => q[1]^2"]
["z[4] => -2*p[1]*q[1] + q[1]^2"]
["z[5] => 2*p[1]*p[2] - 2*p[1]*q[1] - 2*p[2]*q[1] + q[1]^2"]
["z[6] => p[1]*q[2]"]
["z[7] => q[2]^2"]
["z[8] => -2*p[1]*p[2] - 2*p[1]*p[3] - 2*p[1]*p[4] - 2*p[2]*p[3] - 2*p[2]*p[4] - 2*p[3]*q[2] - 2*p[4]*q[2] + q[2]^2"]
["z[9] => -2*p[4]*q[2] + q[2]^2"]
["z[10] => q[1]^2 + 2*q[1]*q[2] + q[2]^2"]
["z[11] => -2*p[1]*q[1] - 2*p[1]*q[2] - 2*p[2]*q[1] - 2*p[2]*q[2] - 2*p[3]*q[1] - 2*p[3]*q[2] - 2*p[4]*q[1] - 2*p[4]*q[2] + q[1]^2 + 2*q[1]*q[2] + q[2]^2"]

julia> G.baikovmatrix
6×6 Matrix{RingElem}:
 0                      …  z[6]
 1//2*t[1]                 1//2*t[2] + 1//2*t[3] + 1//2*t[4] + 1//2*t[5] - z[1] - z[2] - 1//2*z[3] + 1//2*z[5] - z[6] - 1//2*z[7] + 1//2*z[8] + 1//2*z[10] - 1//2*z[11]
 1//2*t[2]                 -1//2*t[1] - 1//2*t[2] - 1//2*t[3] - 1//2*t[4] - 1//2*t[5] - 1//2*z[8] + 1//2*z[9]
 1//2*t[3]                 1//2*z[7] - 1//2*z[9]
 1//2*z[3] - 1//2*z[4]     -1//2*z[3] - 1//2*z[7] + 1//2*z[10]
 z[6]                   …  z[7]

Feynman.computeIBP — Method

computeIBP(G::labeledgraph,Nu::Vector{Int64},cutDeg::Int)

USAGE : computeIBP(G,ν,d);

ASSUME : G is a labeled graph, d is a positive integer and ν is vector of integers correspond to the parent diagram of the integral.

RETURN : A set of simplified IBP identities without double propagators (without performing trimming) .

Feynman.eliminateVariables — Method

eliminateVariables(G::labeledgraph)

USAGE : eliminateVariables(G);

ASSUME : G is a labeled graph.

RETURN : labeled graph with variables of the bounded edges eliminated according to balancing condition.

#Examples

julia> G=simple_graph([1,2,3,4],[(1,3),(1,2),(2,4),(3,4),1,2,3,4]);
julia> G=labelGraph(G,0);
julia> G=eliminateVariables(G);
julia> printLabeledGraph(G);

Graph with 4 vertices and 4 bounded edges 4 unbounded edges
Edge terms:
["(1, 3)=>q[1]", "(1, 2)=>-p[1] - q[1]", "(2, 4)=>-p[1] - p[2] - q[1]", "(3, 4)=>-p[3] + q[1]", "1=>p[1]", "2=>p[2]", "3=>p[3]", "4=>-p[1] - p[2] - p[3]"]
julia> G.elimvar

4-element Vector{QQMPolyRingElem}:
 p[4]
 q[2]
 q[3]
 q[4]

Feynman.feynmanDenominators — Method

feynmanDenominators(G::labeledgraph)

USAGE : feynmanDenominators(G);

ASSUME : $G$ is a labeled graph with the variables of the bounded edges eliminated according to balancing condition. i.e. $G$ is a labeled graph where the function eliminatedVariables applied.

RETURN : ideal containing the propagators in the Feynman integral

#Examples

julia> G=simple_graph([1,2,3,4],[(1,3),(1,2),(2,4),(3,4),1,2,3,4]);
julia> G=labelGraph(G,0);
julia> Gelim=eliminateVariables(G);
julia> feynmanDenominators(Gelim)
Ideal generated by
  q[1]^2
  p[1]^2 + 2*p[1]*q[1] + q[1]^2
  p[1]^2 + 2*p[1]*p[2] + 2*p[1]*q[1] + p[2]^2 + 2*p[2]*q[1] + q[1]^2
  p[3]^2 - 2*p[3]*q[1] + q[1]^2

Feynman.labelGraph — Method

labelGraph(G::simple_graph,ch::Int)

USAGE : labelGraph(G,ch);

ASSUME : G is a graph and ch is either zero or prime.

RETURN : labeled graph with polynomialvariables qi at the bounded edges and functin filed variables pi at the unbounded edges over a prime filed of characteristic ch Initially we it sets the fields Baikovmatrix and elimvar empty.

#Examples

julia> G3=Feynman.simple_graph([1,2,3,4],[(1,3),(1,2),(2,4),(3,4),1,2,3,4]);
julia> G4=Feynman.labelGraph(G3,0);
julia> Feynman.printLabeledGraph(G4);
Graph with 4 vertices and 4 bounded edges 4 unbounded edges
Edge terms:
["(1, 3)=>q[1]", "(1, 2)=>q[2]", "(2, 4)=>q[3]", "(3, 4)=>q[4]", "1=>p[1]", "2=>p[2]", "3=>p[3]", "4=>p[4]"]

Feynman.makePoly — Method

makePoly(n::Int,m::Int)

USAGE : makePoly(m,n);

ASSUME : m and n are positve integers.

RETURN : A polynomial ring with vatiables t[1],...,t[n],z[1],...,z[m] over QQ.

Feynman.printIBP — Method

printIBP(set_IBP::Vector{Vector{}},n::Int64)

USAGE : printIBP(set_IBP,n);

ASSUME : set_IBP is the output of computeIBP

RETURN : It prints first n IBP relations

Feynman.printLabeledGraph — Method

printLabeledGraph(G::labeledgraph)

USAGE : printLabeledGraph(G);

ASSUME : G is a labeled graph.

Theory: This is the print function used in julia to print a labeled graph.

#Examples

julia> var=["x","y","z","p","q","r"];
julia> R, (x,y,z,p,q,r)=polynomial_ring(QQ,var);
julia> G=labeledgraph([1,2,3,4],[(1,3),(1,2),(1,2),(2,4),(3,4),(3,4)],R,var,R,[3,4],R,[[y+1,R(2)] [R(3),r+3]]);
julia> printLabeledGraph(G);

Graph with 4 vertices and 6 edges
Edge terms:
["(1, 3)=>x", "(1, 2)=>y", "(1, 2)=>z", "(2, 4)=>p", "(3, 4)=>q", "(3, 4)=>r"]

Feynman.propagators — Method

propagators(G::labeledgraph)

USAGE: propagators(G);

ASSUME: G is a labeld graph.

RETURN: ideal, containing the denominators in the Feynman integral.

#Examples

julia> G=simple_graph([1,2,3,4,5,6],[(1,2),(3,6),(4,5),(1,6),(2,3),(5,6),(3,4),1,2,5,4]);

julia> G=labelGraph(G,0);

julia> Gelim=removeElimVars(Gelim);

julia> Gelim=removeElimVars(Gelim);

julia> propagators(Gelim)
Ideal generated by
  q[1]^2
  q[2]^2
  2*p[1]*p[3] + 2*p[1]*q[1] - 2*p[1]*q[2] + 2*p[3]*q[1] - 2*p[3]*q[2] + q[1]^2 - 2*q[1]*q[2] + q[2]^2
  2*p[1]*q[1] + q[1]^2
  -2*p[2]*q[1] + q[1]^2
  2*p[1]*q[1] - 2*p[1]*q[2] + q[1]^2 - 2*q[1]*q[2] + q[2]^2
  -2*p[2]*q[1] + 2*p[2]*q[2] + q[1]^2 - 2*q[1]*q[2] + q[2]^2

Feynman.removeElimVars — Method

removeElimVars(G::labeledgraph)

USAGE : removeElimVars(G);

ASSUME : G is a labled graph.

RETURN : Removes the variables from G.elimvars. This key is generated by the procedure eliminatedVariables.

#Examples

julia> G=simple_graph([1,2,3,4,5,6],[(1,2),(3,6),(4,5),(1,6),(2,3),(5,6),(3,4),1,2,5,4]);

julia> G=labelGraph(G,0);

julia> Gelim=eliminateVariables(G);

julia> G=removeElimVars(Gelim);
QQMPolyRingElem[p[1], p[2], p[3], p[4], q[1], q[2]]
julia> G.elimvar
Any[]

Feynman.removeParameter — Method

removeParameter(P::Ring,l::Vector)

USAGE : removeParameter(R,l);

ASSUME : R is a polynomial ring.

RETURN : polynomial ring with the parameters at indeces in l removed.

#Examples

julia> R,v,w=polynomial_ring(QQ,"p"=>(1:5),"q"=>(1:6));
julia> I=ideal(R,[v[1],v[2],v[3],v[4],v[5],w[1]]);
julia> u=complement_of_prime_ideal(I);
julia> S,iso=localization(R,u);
julia> S

julia> S
Localization
  of multivariate polynomial ring in 11 variables p[1], p[2], p[3], p[4], ..., q[6]
    over rational field
  at complement of prime ideal (p[1], p[2], p[3], p[4], p[5], q[1])

  julia> removeParameter(S,[2]) 
Localization
  of multivariate polynomial ring in 10 variables p[1], p[3], p[4], p[5], ..., q[6]
    over rational field
  at complement of prime ideal (p[1], p[3], p[4], p[5], q[1])

Feynman.removeVariable — Method

removeVariable(R::Ring,l::Vector)

USAGE: removeVariable(G,l);

ASSUME: R is a polynomial ring.

RERUTN: polynomial ring with the vaiables at indeces given in l removed.

#Examples

julia> R,v=polynomial_ring(QQ,"p"=>(1:3));
julia> Feynman.removeVariable(R,[2])
Multivariate polynomial ring in 2 variables p[1], p[3]
  over rational field

Feynman.removeVariableLocal — Method

removeVariableLocal(P::Ring,l::Vector)

USAGE : removeVariableLocal(P,l);

ASSUME : P is a local ring locaized by the maximal ideal generated by parameters.

RETURN : local ring where the variables at indeces in l removed.

Feynman.substituteGraph — Method

substituteGraph(G::labeledgraph,a::RingElement,b::RingElement)

USAGE :substituteGraph(G,a,b)

ASSUME : G is a labeled graph

RETURN :a labelled graph with labelling where each 'a' is replaced 'b'