Overview

Feynman.Feynman

Feynman.ISP

Feynman.balancingIdeal

Feynman.computeIBP

Feynman.eliminateVariables

Feynman.feynmanDenominators

Feynman.makePoly

Feynman.printIBP

Feynman.propagators

Feynman.removeElimVars

Feynman.removeParameter

Feynman.removeVariable

Feynman.removeVariableLocal

Feynman.Feynman — Module

Feynman is a package for computing integration-by-part identities (IBPs) of a Feynman Integral associated to Feynman graph using module intesecation method. 
This package also provides an interface of Oscar to use the packages NeatIBP developed using Singular and GPI-Space.

Feynman.ISP — Method

ISP(G::labeledgraph)

USAGE: ISP(G); ASSUME: G is a labeled graph. RETURN: idal containing the irreducible scalar products(ISPs), that is, those scalar product which are not linearly dependent on the propagators.

#Examples

julia> G=simple_graph([1,2,3,4,5,6],[(1,2),(3,6),(4,5),(1,6),(2,3),(5,6),(3,4),1,2,5,4]);

julia> G=labelGraph(G,0);

julia> Gelim=eliminateVariables(G);

julia> ISP(Gelim)
Ideal generated by
  p[3]*q[1]
  p[1]*q[2]

Feynman.balancingIdeal — Method

balancingIdeal(G::labeledgraph)

USAGE: balancingIdeal(G) ASSUME: G is a labeled graph. RETURN: Ideal of balancing condition of the graph. i.e Ideal generated by the relation of the momentums which are obtained by applying momentum conservation law to external mementa, and at each vertex; This is an ideal of the ring G.over.

#Examples

julia> G=simple_graph([1,2,3,4],[(1,3),(1,2),(2,4),(3,4),1,2,3,4]);
julia> G=labelGraph(G,0);
julia> balancingIdeal(G)

Ideal generated by
p[1] + p[2] + p[3] + p[4]
p[1] + q[1] + q[2]
p[2] - q[2] + q[3]
p[3] - q[1] + q[4]
p[4] - q[3] - q[4]

Feynman.computeIBP — Method

computeIBP(G::labeledgraph,Nu::Vector{Int64},cutDeg::Int)

USAGE: computeIBP(G,ν,d); ASSUME: G is a labeled graph, d is a positive integer and ν is vector of integers correspond to the parent diagram of the integral. RETURN: A set of simplified IBP identities without double propagators (without performing trimming) .

Feynman.eliminateVariables — Method

eliminateVariables(G::labeledgraph)

USAGE: eliminateVariables(G) ASSUME: G is a labeled graph. RETURN: labeled graph with variables of the bounded edges eliminated according to balancing condition.

#Examples

julia> G=simple_graph([1,2,3,4],[(1,3),(1,2),(2,4),(3,4),1,2,3,4]);
julia> G=labelGraph(G,0);
julia> G=eliminateVariables(G);
julia> printLabeledGraph(G);

Graph with 4 vertices and 4 bounded edges 4 unbounded edges
Edge terms:
["(1, 3)=>q[1]", "(1, 2)=>-p[1] - q[1]", "(2, 4)=>-p[1] - p[2] - q[1]", "(3, 4)=>-p[3] + q[1]", "1=>p[1]", "2=>p[2]", "3=>p[3]", "4=>-p[1] - p[2] - p[3]"]
julia> G.elimvar

4-element Vector{QQMPolyRingElem}:
 p[4]
 q[2]
 q[3]
 q[4]

Feynman.feynmanDenominators — Method

feynmanDenominators(G::labeledgraph)

USAGE: feynmanDenominators(G); ASSUME: G is a labeled graph with the variables of the bounded edges eliminated according to balancing condition. i.e. G is a labeled graph where the function eliminatedVariables applied. RETURN: ideal containing the propagators in the Feynman integral

#Examples

julia> G=simple_graph([1,2,3,4],[(1,3),(1,2),(2,4),(3,4),1,2,3,4]);
julia> G=labelGraph(G,0);
julia> Gelim=eliminateVariables(G);
julia> feynmanDenominators(Gelim)
Ideal generated by
  q[1]^2
  p[1]^2 + 2*p[1]*q[1] + q[1]^2
  p[1]^2 + 2*p[1]*p[2] + 2*p[1]*q[1] + p[2]^2 + 2*p[2]*q[1] + q[1]^2
  p[3]^2 - 2*p[3]*q[1] + q[1]^2

Feynman.makePoly — Method

makePoly(n::Int,m::Int)

USAGE: makePoly(m,n); ASSUME: m and n are positve integers. RETURN: A polynomial ring with vatiables t[1],...,t[n],z[1],...,z[m] over QQ.

Feynman.printIBP — Method

printIBP(setIBP::Vector{Vector{}},n::Int64) USAGE: printIBP(singleIBP,n);

Feynman.propagators — Method

propagators(G::labeledgraph)

USAGE: propagators(G) ASSUME: G is a labeld graph. RETURN: ideal, containing the denominators in the Feynman integral.

#Examples

julia> G=simple_graph([1,2,3,4,5,6],[(1,2),(3,6),(4,5),(1,6),(2,3),(5,6),(3,4),1,2,5,4]);

julia> G=labelGraph(G,0);

julia> Gelim=removeElimVars(Gelim);

julia> Gelim=removeElimVars(Gelim);

julia> propagators(Gelim)
Ideal generated by
  q[1]^2
  q[2]^2
  2*p[1]*p[3] + 2*p[1]*q[1] - 2*p[1]*q[2] + 2*p[3]*q[1] - 2*p[3]*q[2] + q[1]^2 - 2*q[1]*q[2] + q[2]^2
  2*p[1]*q[1] + q[1]^2
  -2*p[2]*q[1] + q[1]^2
  2*p[1]*q[1] - 2*p[1]*q[2] + q[1]^2 - 2*q[1]*q[2] + q[2]^2
  -2*p[2]*q[1] + 2*p[2]*q[2] + q[1]^2 - 2*q[1]*q[2] + q[2]^2

Feynman.removeElimVars — Method

removeElimVars(G::labeledgraph)

USAGE: removeElimVars(G); ASSUME: G is a labled graph. RETURN: Removes the variables from G.elimvars. This key is generated by the procedure eliminatedVariables.

#Examples ```julia julia> G=simple_graph([1,2,3,4,5,6],[(1,2),(3,6),(4,5),(1,6),(2,3),(5,6),(3,4),1,2,5,4]);

julia> G=labelGraph(G,0);

julia> Gelim=eliminateVariables(G);

julia> G=removeElimVars(Gelim); QQMPolyRingElem[p[1], p[2], p[3], p[4], q[1], q[2]] julia> G.elimvar Any[]```

Feynman.removeParameter — Method

removeParameter(P::Ring,l::Vector)

USAGE: removeParameter(R,l); ASSUME: R is a polynomial ring. RETURN: polynomial ring with the parameters at indeces in l removed.

#Examples

julia> R,v,w=polynomial_ring(QQ,"p"=>(1:5),"q"=>(1:6));
julia> I=ideal(R,[v[1],v[2],v[3],v[4],v[5],w[1]]);
julia> u=complement_of_prime_ideal(I);
julia> S,iso=localization(R,u);
julia> S

julia> S
Localization
  of multivariate polynomial ring in 11 variables p[1], p[2], p[3], p[4], ..., q[6]
    over rational field
  at complement of prime ideal (p[1], p[2], p[3], p[4], p[5], q[1])

  julia> removeParameter(S,[2]) 
Localization
  of multivariate polynomial ring in 10 variables p[1], p[3], p[4], p[5], ..., q[6]
    over rational field
  at complement of prime ideal (p[1], p[3], p[4], p[5], q[1])

Feynman.removeVariable — Method

removeVariable(R::Ring,l::Vector)

USAGE: removeVariable(G,l) ASSUME: R is a polynomial ring. RERUTN: polynomial ring with the vaiables at indeces given in l removed.

#Examples

julia> R,v=polynomial_ring(QQ,"p"=>(1:3));
julia> Feynman.removeVariable(R,[2])
Multivariate polynomial ring in 2 variables p[1], p[3]
  over rational field

Feynman.removeVariableLocal — Method

removeVariableLocal(P::Ring,l::Vector)

USAGE: removeVariableLocal(P,l); ASSUME: P is a local ring locaized by the maximal ideal generated by parameters. RETURN: local ring where the variables at indeces in l removed.